Make footer, Again make header and First-section
footer make this time I not used any position property as discussed in the project instead I use flexbox to justify and align to div.

header was not enough as it is not well made so I remake it.
Similarly with First-section as it confused me a lot, I think it is due to padding but is not I removed padding and suddenly so I understand that div take space of which elements need otherwise you have to specify it to apply flexbox such as justify-content: spacec-between .